feat: input animation 추가, join page 구현
