+++ src/app/@modal/auth/join/components/joinForm/index.ts
@@ -1,0 +1,3 @@
+export * from "./JoinForm";
+export * from "./joinForm.variants";
+
